Move tests to cursorBack and cursorForward
